requirements.txt
Flask
Jinja2
numpy
pandas
plotly==5.5.0
python-dotenv
pyyaml
selenium